second_service/requirements.txt
django==4.2.13
djangorestframework==3.15.2
psycopg2==2.9.9
gunicorn==22.0.0
celery==5.4.0
gym==0.26.2
